Clean up XCTAssertEqual use which no longer requires precise type match

+++ Sample/Tests/ViewControllerTests-PlainXCTest.m
@@ -41,7 +41,7 @@
     NSArray *touchUpActions = [sut.showAlertButton actionsForTarget:sut
                                                     forControlEvent:UIControlEventTouchUpInside];
 
-    XCTAssertEqual(touchUpActions.count, (NSUInteger)1);
+    XCTAssertEqual(touchUpActions.count, 1);
     XCTAssertEqualObjects(touchUpActions[0], @"showAlert:");
 }
 
@@ -52,7 +52,7 @@
     NSArray *touchUpActions = [sut.showActionSheetButton actionsForTarget:sut
                                                           forControlEvent:UIControlEventTouchUpInside];
 
-    XCTAssertEqual(touchUpActions.count, (NSUInteger)1);
+    XCTAssertEqual(touchUpActions.count, 1);
     XCTAssertEqualObjects(touchUpActions[0], @"showActionSheet:");
 }
 
@@ -62,7 +62,7 @@
 
     [sut showAlert:nil];
     
-    XCTAssertEqual(alertVerifier.presentedCount, (NSUInteger)1);
+    XCTAssertEqual(alertVerifier.presentedCount, 1);
 }
 
 - (void)testShowActionSheet_ShouldPresentModal
@@ -71,7 +71,7 @@
 
     [sut showActionSheet:nil];
 
-    XCTAssertEqual(alertVerifier.presentedCount, (NSUInteger)1);
+    XCTAssertEqual(alertVerifier.presentedCount, 1);
 }
 
 - (void)testShowAlert_ShouldPreferAlert
@@ -152,7 +152,7 @@
 
     [sut showAlert:nil];
 
-    XCTAssertEqual(alertVerifier.actionTitles.count, (NSUInteger)4);
+    XCTAssertEqual(alertVerifier.actionTitles.count, 4);
     XCTAssertEqualObjects(alertVerifier.actionTitles[0], @"Default");
     XCTAssertEqualObjects(alertVerifier.actionTitles[1], @"Cancel");
     XCTAssertEqualObjects(alertVerifier.actionTitles[2], @"Destroy");
@@ -165,7 +165,7 @@
 
     [sut showActionSheet:nil];
 
-    XCTAssertEqual(alertVerifier.actionTitles.count, (NSUInteger)4);
+    XCTAssertEqual(alertVerifier.actionTitles.count, 4);
     XCTAssertEqualObjects(alertVerifier.actionTitles[0], @"Default");
     XCTAssertEqualObjects(alertVerifier.actionTitles[1], @"Cancel");
     XCTAssertEqualObjects(alertVerifier.actionTitles[2], @"Destroy");
